requirements.txt
djangorestframework
# # # django==3.0.5
# # pymongo==3.10.1
# # djongo==1.3.3
# # mongoengine
# # django-allauth
# # django-sslserver==0.22
# PyJWT
# cryptography
# cffi
# Pillow
# pip
# wheel
# py
# bson
# pymongo
# Django
# django-mptt
# pytz
# Babel
# Unidecode
# six
# sqlparse
# asgiref
# docutils
# djongo
# dataclasses
# django-allauth
# requests
# certifi
# chardet
# oauthlib
# idna
# urllib3
# simplejson
# setuptools
# mongoengine
# python-dateutil
# django-debug-toolbar
# django-filter
# djangorestframework
# Markdown
# jsonfield
# djangorestframework-jwt